Extract camera settings into its own component
